Made it a simple console app with all code in main
